add robustness measures

+++ src/python/typ.py
@@ -1,7 +1,7 @@
 ## Type Aliases ##
 Edge = tuple[int, int]
 KCut = tuple[int, list[int]]  # cuts under KaHIP format
-EdgeDict = dict[Edge, int]  # common edge dict
+EdgeDict = dict[Edge, float]  # common edge dict
 EdgeDict3 = dict[tuple[int, int, int], int]  # common edge dict
 EdgeDictStr = dict[str, int]  # edge dict after json import
 Cuts = dict[
